Removed test images
