+++ test/lib/operation.go
@@ -23,7 +23,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	pkgTest "knative.dev/pkg/test"
 
-	"knative.dev/eventing/test/lib/cloudevents"
 	"knative.dev/eventing/test/lib/duck"
 	"knative.dev/eventing/test/lib/resources"
 )
@@ -45,40 +44,6 @@ func (c *Client) LabelNamespace(labels map[string]string) error {
 	return err
 }
 
-// SendFakeEventToAddressableOrFail will send the given event to the given Addressable.
-// Deprecated: you should use SendEventToAddressable
-func (c *Client) SendFakeEventToAddressableOrFail(
-	senderName,
-	addressableName string,
-	typemeta *metav1.TypeMeta,
-	event *cloudevents.CloudEvent,
-) {
-	uri, err := c.GetAddressableURI(addressableName, typemeta)
-	if err != nil {
-		c.T.Fatalf("Failed to get the URI for %v-%s", typemeta, addressableName)
-	}
-	if err = c.sendFakeEventToAddress(senderName, uri, event); err != nil {
-		c.T.Fatalf("Failed to send event %v to %s: %v", event, uri, err)
-	}
-}
-
-// SendFakeEventWithTracingToAddressableOrFail will send the given event with tracing to the given Addressable.
-// Deprecated: you should use SendEventToAddressable
-func (c *Client) SendFakeEventWithTracingToAddressableOrFail(
-	senderName,
-	addressableName string,
-	typemeta *metav1.TypeMeta,
-	event *cloudevents.CloudEvent,
-) {
-	uri, err := c.GetAddressableURI(addressableName, typemeta)
-	if err != nil {
-		c.T.Fatalf("Failed to get the URI for %v-%s", typemeta, addressableName)
-	}
-	if err = c.sendFakeEventWithTracingToAddress(senderName, uri, event); err != nil {
-		c.T.Fatalf("Failed to send event %v with tracing to %s: %v", event, uri, err)
-	}
-}
-
 // GetAddressableURI returns the URI of the addressable resource.
 // To use this function, the given resource must have implemented the Addressable duck-type.
 func (c *Client) GetAddressableURI(addressableName string, typeMeta *metav1.TypeMeta) (string, error) {
@@ -89,44 +54,6 @@ func (c *Client) GetAddressableURI(addressableName string, typeMeta *metav1.Type
 		return "", err
 	}
 	return u.String(), nil
-}
-
-// sendFakeEventToAddress will create a sender pod, which will send the given event to the given url.
-// Deprecated: you should use SendEvent
-func (c *Client) sendFakeEventToAddress(
-	senderName string,
-	uri string,
-	event *cloudevents.CloudEvent,
-) error {
-	namespace := c.Namespace
-	pod, err := resources.EventSenderPod(senderName, uri, event)
-	if err != nil {
-		return err
-	}
-	c.CreatePodOrFail(pod)
-	if err := pkgTest.WaitForPodRunning(c.Kube, senderName, namespace); err != nil {
-		return err
-	}
-	return nil
-}
-
-// sendFakeEventWithTracingToAddress will create a sender pod, which will send the given event with tracing to the given url.
-// Deprecated: you should use SendEvent
-func (c *Client) sendFakeEventWithTracingToAddress(
-	senderName string,
-	uri string,
-	event *cloudevents.CloudEvent,
-) error {
-	namespace := c.Namespace
-	pod, err := resources.EventSenderTracingPod(senderName, uri, event)
-	if err != nil {
-		return err
-	}
-	c.CreatePodOrFail(pod)
-	if err := pkgTest.WaitForPodRunning(c.Kube, senderName, namespace); err != nil {
-		return err
-	}
-	return nil
 }
 
 // WaitForResourceReadyOrFail waits for the resource to become ready or fail.
