+++ test/lib/resources/kube.go
@@ -19,7 +19,6 @@ package resources
 // This file contains functions that construct common Kubernetes resources.
 
 import (
-	"encoding/json"
 	"fmt"
 
 	v1 "knative.dev/pkg/apis/duck/v1"
@@ -30,8 +29,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	pkgTest "knative.dev/pkg/test"
-
-	cetest "knative.dev/eventing/test/lib/cloudevents"
 )
 
 // PodOption enables further configuration of a Pod.
@@ -39,79 +36,6 @@ type PodOption func(*corev1.Pod)
 
 // Option enables further configuration of a Role.
 type RoleOption func(*rbacv1.Role)
-
-// EventSenderPod creates a Pod that sends a single event to the given address.
-// Deprecated: use sender.EventSenderPod
-func EventSenderPod(name string, sink string, event *cetest.CloudEvent) (*corev1.Pod, error) {
-	return eventSenderPodImage("sendevents", name, sink, event, false)
-}
-
-// EventSenderTracingPod creates a Pod that sends a single event to the given address.
-// Deprecated: use sender.EventSenderPod
-func EventSenderTracingPod(name string, sink string, event *cetest.CloudEvent) (*corev1.Pod, error) {
-	return eventSenderPodImage("sendevents", name, sink, event, true)
-}
-
-// Deprecated: use sender.EventSenderPod
-func eventSenderPodImage(imageName string, name string, sink string, event *cetest.CloudEvent, addTracing bool) (*corev1.Pod, error) {
-	if event.Encoding == "" {
-		event.Encoding = cetest.DefaultEncoding
-	}
-	eventExtensionsBytes, err := json.Marshal(event.Extensions)
-	eventExtensions := string(eventExtensionsBytes)
-	if err != nil {
-		return nil, fmt.Errorf("encountered error when we marshall cloud event extensions %v", err)
-	}
-
-	args := []string{
-		"-event-id",
-		event.ID,
-		"-event-type",
-		event.Type,
-		"-event-source",
-		event.Source.String(),
-		"-event-extensions",
-		eventExtensions,
-		"-event-data",
-		event.Data,
-		"-event-encoding",
-		event.Encoding,
-		"-sink",
-		sink,
-	}
-	if addTracing {
-		args = append(args, "-add-tracing")
-	}
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: corev1.PodSpec{
-			Containers: []corev1.Container{{
-				Name:            imageName,
-				Image:           pkgTest.ImagePath(imageName),
-				ImagePullPolicy: corev1.PullAlways,
-				Args:            args,
-			}},
-			// Never restart the event sender Pod.
-			RestartPolicy: corev1.RestartPolicyNever,
-		},
-	}, nil
-}
-
-// EventLoggerPod creates a Pod that logs events received.
-// Deprecated: This test image is gonna be removed soon and you should use EventRecordPod.
-//             Look at recordevents.StartEventRecordOrFail for more info
-func EventLoggerPod(name string) *corev1.Pod {
-	return eventLoggerPod("logevents", name)
-}
-
-// EventDetailsPod creates a Pod that validates events received and log details about events.
-// Deprecated: This test image is gonna be removed soon and you should use EventRecordPod.
-//             Look at recordevents.StartEventRecordOrFail for more info
-func EventDetailsPod(name string) *corev1.Pod {
-	return eventLoggerPod("eventdetails", name)
-}
 
 // EventRecordPod creates a Pod that stores received events for test retrieval.
 func EventRecordPod(name string) *corev1.Pod {
@@ -129,35 +53,6 @@ func eventLoggerPod(imageName string, name string) *corev1.Pod {
 				Name:            imageName,
 				Image:           pkgTest.ImagePath(imageName),
 				ImagePullPolicy: corev1.PullAlways,
-			}},
-			RestartPolicy: corev1.RestartPolicyAlways,
-		},
-	}
-}
-
-// DeprecatedEventTransformationPod creates a Pod that transforms events received.
-// Deprecated: Use EventTransformationPod
-// TODO(nlopezgi): remove once other tests that use sdk1 and depend on this method are migrated.
-func DeprecatedEventTransformationPod(name string, event *cetest.CloudEvent) *corev1.Pod {
-	const imageName = "transformevents"
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   name,
-			Labels: map[string]string{"e2etest": string(uuid.NewUUID())},
-		},
-		Spec: corev1.PodSpec{
-			Containers: []corev1.Container{{
-				Name:            imageName,
-				Image:           pkgTest.ImagePath(imageName),
-				ImagePullPolicy: corev1.PullAlways,
-				Args: []string{
-					"-event-type",
-					event.Type,
-					"-event-source",
-					event.Source.String(),
-					"-event-data",
-					event.Data,
-				},
 			}},
 			RestartPolicy: corev1.RestartPolicyAlways,
 		},
