--- test/test_images/eventdetails/README.md
@@ -1,13 +1,0 @@
-# Event Details Image
-
-Validates received events and print details for conformance testing.
-
-## Development
-
-Can be run without k8s for quick dev testing of code:
-
-```go
-go run knative.dev/eventing/test/test_images/eventdetails
-```
-
-And use [sendevents](../sendevents) with to get events.
