--- test/test_images/logevents/main.go
@@ -1,49 +1,0 @@
-/*
-Copyright 2019 The Knative Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"context"
-	"log"
-
-	cloudevents "github.com/cloudevents/sdk-go"
-	"go.uber.org/zap"
-
-	"knative.dev/eventing/pkg/kncloudevents"
-	"knative.dev/eventing/pkg/tracing"
-)
-
-func handler(event cloudevents.Event) {
-	if err := event.Validate(); err == nil {
-		log.Printf("%s", event.String())
-	} else {
-		log.Printf("error validating the event: %v", err)
-	}
-}
-
-func main() {
-	logger, _ := zap.NewDevelopment()
-	if err := tracing.SetupStaticPublishing(logger.Sugar(), "", tracing.AlwaysSample); err != nil {
-		log.Fatalf("Unable to setup trace publishing: %v", err)
-	}
-	c, err := kncloudevents.NewDefaultClient()
-	if err != nil {
-		log.Fatalf("failed to create client, %v", err)
-	}
-
-	log.Fatalf("failed to start receiver: %s", c.StartReceiver(context.Background(), handler))
-}
