--- test/test_images/logevents/pod.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: logevents
-spec:
-  containers:
-    - name: logevents
-      image: ko://knative.dev/eventing/test/test_images/logevents
-
